extracted code in Daily to separate component
fail was getting too long due to static html on top of dynamic data

will make it easier to implement loading state
