+++ client/src/components/index.ts
@@ -1,5 +1,6 @@
 export * from './Access';
 export * from './DailyMedication';
+export * from './DailySummary';
 export * from './Header';
 export * from './Loading';
 export * from './Medication';
